I mean, it works now
Added runCommand to replace std::system, which required use of temporary
files and IO streams to bring the file back in. Thanks Block.
Uses an input string as a command, uses a FILE* to take it's output and
forwards the FILE* to a lambda in the form of a std::string with a
length of the buffer size

Added Block's codepointToUTF8 function, since working with UTF-8 in any
other way is tiring.